#33 테스트 커밋

+++ Ami/src/main/java/com/cnu/ami/device/server/dao/entity/ServerEntity.java
@@ -16,7 +16,7 @@ public class ServerEntity {
 
 	@Id
 	@Column(name = "SSEQ")
-	private int SSEQ;
+	private long SSEQ;
 
 	@Column(name = "SNAME")
 	private String SNAME;
