+++ Ami/src/main/java/com/cnu/ami/device/server/models/ServerListVO.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 public class ServerListVO {
 
-	private int serverSeq;
+	private long serverSeq;
 	private String serverName;
 	private String model;
 	private String ip;
